Fix: Correct import path for useAuth after refactoring
I've fixed a runtime error (`Uncaught SyntaxError: The requested module '/src/shared/contexts/AuthContext.jsx' does not provide an export named 'useAuth'`) that was occurring after the refactoring of the contexts.

The `useAuth` hook was moved from `AuthContext.jsx` to `authHooks.js`. This update ensures all imports of `useAuth` now point to the new correct location (`frontend/src/shared/contexts/authHooks.js`).

Affected files:
- `frontend/src/shared/components/layout/NavbarAdmin.jsx`
- `frontend/src/features/auth/pages/LoginPage.jsx`
- `frontend/src/shared/components/auth/PrivateRoute.jsx`

Additionally, I verified that the imports for `CartContext` (which was moved to `cartContextValue.js`) were correct and didn't need further changes.

I ran ESLint to confirm there are no new linting issues.
